ScenariosSelector n'était plus utilisé
